Refactor code to be able to use Dyson purifier and 360 Eye devices
